requirements.txt
black>=25.1.0
gradio>=4.44.1
ipykernel>=6.29.5
litellm>=1.72.0
openai>=1.83.0
python-dotenv>=1.1.0
ruff>=0.11.12